Report the chain of properties from which the value of a finalized `Property` has come from, when the value is missing.
