+++ subprojects/model-core/src/main/java/org/gradle/api/internal/provider/AbstractMinimalProvider.java
@@ -21,7 +21,6 @@ import org.gradle.api.Task;
 import org.gradle.api.Transformer;
 import org.gradle.api.internal.tasks.TaskDependencyResolveContext;
 import org.gradle.api.provider.Provider;
-import org.gradle.internal.Cast;
 import org.gradle.internal.Describables;
 import org.gradle.internal.DisplayName;
 import org.gradle.internal.logging.text.TreeFormatter;
@@ -158,8 +157,7 @@ public abstract class AbstractMinimalProvider<T> implements ProviderInternal<T>,
 
     @Override
     public ScalarSupplier<T> withFinalValue() {
-        T value = getOrNull();
-        return Providers.nullableValue(value);
+        return Providers.nullableValue(calculateValue());
     }
 
     @Override
@@ -188,151 +186,4 @@ public abstract class AbstractMinimalProvider<T> implements ProviderInternal<T>,
         return ManagedFactories.ProviderManagedFactory.FACTORY_ID;
     }
 
-    private static class FlatMapProvider<S, T> extends AbstractMinimalProvider<S> {
-        private final ProviderInternal<? extends T> provider;
-        private final Transformer<? extends Provider<? extends S>, ? super T> transformer;
-
-        FlatMapProvider(ProviderInternal<? extends T> provider, Transformer<? extends Provider<? extends S>, ? super T> transformer) {
-            this.provider = provider;
-            this.transformer = transformer;
-        }
-
-        @Nullable
-        @Override
-        public Class<S> getType() {
-            return null;
-        }
-
-        @Override
-        public boolean isPresent() {
-            T value = provider.getOrNull();
-            if (value == null) {
-                return false;
-            }
-            return map(value).isPresent();
-        }
-
-        @Override
-        protected Value<? extends S> calculateOwnValue() {
-            Value<? extends T> value = provider.calculateValue();
-            if (value.isMissing()) {
-                return value.asType();
-            }
-            return map(value.get()).calculateValue();
-        }
-
-        private ProviderInternal<? extends S> map(T value) {
-            Provider<? extends S> result = transformer.transform(value);
-            if (result == null) {
-                return Providers.notDefined();
-            }
-            return Providers.internal(result);
-        }
-
-        @Override
-        public boolean maybeVisitBuildDependencies(TaskDependencyResolveContext context) {
-            return Providers.internal(map(provider.get())).maybeVisitBuildDependencies(context);
-        }
-
-        @Override
-        public String toString() {
-            return "flatmap(" + provider + ")";
-        }
-    }
-
-    private static class OrElseFixedValueProvider<T> extends AbstractProviderWithValue<T> {
-        private final ProviderInternal<T> provider;
-        private final T value;
-
-        public OrElseFixedValueProvider(ProviderInternal<T> provider, T value) {
-            this.provider = provider;
-            this.value = value;
-        }
-
-        @Nullable
-        @Override
-        public Class<T> getType() {
-            return provider.getType();
-        }
-
-        @Override
-        public boolean maybeVisitBuildDependencies(TaskDependencyResolveContext context) {
-            if (provider.isPresent()) {
-                return provider.maybeVisitBuildDependencies(context);
-            } else {
-                return super.maybeVisitBuildDependencies(context);
-            }
-        }
-
-        @Override
-        public T get() {
-            T value = provider.getOrNull();
-            return value != null ? value : this.value;
-        }
-    }
-
-    private static class OrElseProvider<T> extends AbstractMinimalProvider<T> {
-        private final ProviderInternal<T> left;
-        private final ProviderInternal<? extends T> right;
-
-        public OrElseProvider(ProviderInternal<T> left, ProviderInternal<? extends T> right) {
-            this.left = left;
-            this.right = right;
-        }
-
-        @Nullable
-        @Override
-        public Class<T> getType() {
-            return left.getType();
-        }
-
-        @Override
-        public boolean maybeVisitBuildDependencies(TaskDependencyResolveContext context) {
-            if (left.isPresent()) {
-                return left.maybeVisitBuildDependencies(context);
-            } else {
-                return right.maybeVisitBuildDependencies(context);
-            }
-        }
-
-        @Override
-        public boolean isPresent() {
-            return left.isPresent() || right.isPresent();
-        }
-
-        @Override
-        protected Value<? extends T> calculateOwnValue() {
-            Value<? extends T> leftValue = left.calculateValue();
-            if (!leftValue.isMissing()) {
-                return leftValue;
-            }
-            Value<? extends T> rightValue = right.calculateValue();
-            if (!rightValue.isMissing()) {
-                return rightValue;
-            }
-            return leftValue.addPathsFrom(rightValue);
-        }
-    }
-
-    private static class TypeSanitizingProvider<T> extends AbstractMappingProvider<T, T> {
-        private final DisplayName owner;
-        private final ValueSanitizer<? super T> sanitizer;
-        private final Class<? super T> targetType;
-
-        public TypeSanitizingProvider(DisplayName owner, ValueSanitizer<? super T> sanitizer, Class<? super T> targetType, ProviderInternal<? extends T> delegate) {
-            super(Cast.uncheckedNonnullCast(targetType), delegate);
-            this.owner = owner;
-            this.sanitizer = sanitizer;
-            this.targetType = targetType;
-        }
-
-        @Override
-        protected T mapValue(T v) {
-            v = Cast.uncheckedCast(sanitizer.sanitize(v));
-            if (targetType.isInstance(v)) {
-                return v;
-            }
-            throw new IllegalArgumentException(String.format("Cannot get the value of %s of type %s as the provider associated with this property returned a value of type %s.", owner.getDisplayName(), targetType.getName(), v.getClass().getName()));
-        }
-    }
 }
