+++ subprojects/model-core/src/main/java/org/gradle/api/internal/provider/DefaultProviderFactory.java
@@ -51,7 +51,7 @@ public class DefaultProviderFactory implements ProviderFactory {
         if (value == null) {
             throw new IllegalArgumentException("Value cannot be null");
         }
-        return new DefaultProvider<T>(value);
+        return new DefaultProvider<>(value);
     }
 
     @Override
